Add test for pass

+++ accel/tests/launch_async/mut_ref_fail.rs
@@ -1,7 +1,7 @@
 use accel::*;
 
 #[kernel]
-unsafe fn add(a: *const f32, b: *const f32, c: *mut f32, n: usize) {
+unsafe fn add(a: *const u32, b: *const u32, c: *mut u32, n: usize) {
     let i = accel_core::index();
     if (i as usize) < n {
         *c.offset(i) = *a.offset(i) + *b.offset(i);
@@ -13,13 +13,13 @@ async fn main() -> error::Result<()> {
     let device = Device::nth(0)?;
     let ctx = device.create_context();
     let n = 16;
-    let mut a = DeviceMemory::<f32>::zeros(&ctx, n);
-    let mut b = DeviceMemory::<f32>::zeros(&ctx, n);
-    let mut c = DeviceMemory::<f32>::zeros(&ctx, n);
+    let mut a = DeviceMemory::<u32>::zeros(&ctx, n);
+    let mut b = DeviceMemory::<u32>::zeros(&ctx, n);
+    let mut c = DeviceMemory::<u32>::zeros(&ctx, n);
 
     for i in 0..n {
-        a[i] = i as f32;
-        b[i] = 2.0 * i as f32;
+        a[i] = i as u32;
+        b[i] = 2 * i as u32;
     }
 
     let md = add::Module::new(&ctx)?;
